+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 - add: new context menu item: create reference addons based on selected addons
+- add: new tool: File Name Fixer (this tool might help you solve the problem of missing addon files)
+- add: web proxy credential settings (issue #19)
 - improve: import of addons
 - improve: addon problem check system
 - improve: the backup file won't be created if it's the same as the latest one
-- add: new tool: File Name Fixer (this tool might help you solve the problem of missing addon files)
 - improve: GUI appearance
 - bugfixes
 # v0.7.0